Restructure spidermon.contrib.scrapy.monitors
